chore: sync ckb-indexer json struct
