+++ src/rpc/ckb_light_client.rs
@@ -7,54 +7,14 @@ use ckb_jsonrpc_types::{
 use ckb_types::H256;
 
 pub use crate::rpc::ckb_indexer::{
-    Cell, CellsCapacity, Order, Pagination, ScriptType, SearchKeyFilter,
+    Cell, CellType, CellsCapacity, Order, Pagination, ScriptType, SearchKey, SearchKeyFilter,
 };
-use crate::traits::{CellQueryOptions, ValueRangeOption};
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct ScriptStatus {
     pub script: Script,
     pub script_type: ScriptType,
     pub block_number: BlockNumber,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct SearchKey {
-    pub script: Script,
-    pub script_type: ScriptType,
-    pub filter: Option<SearchKeyFilter>,
-    pub with_data: Option<bool>,
-    pub group_by_transaction: Option<bool>,
-}
-
-impl From<CellQueryOptions> for SearchKey {
-    fn from(opts: CellQueryOptions) -> SearchKey {
-        let convert_range =
-            |range: ValueRangeOption| [Uint64::from(range.start), Uint64::from(range.end)];
-        let filter = if opts.secondary_script.is_none()
-            && opts.script_len_range.is_none()
-            && opts.data_len_range.is_none()
-            && opts.capacity_range.is_none()
-            && opts.block_range.is_none()
-        {
-            None
-        } else {
-            Some(SearchKeyFilter {
-                script: opts.secondary_script.map(|v| v.into()),
-                script_len_range: opts.script_len_range.map(convert_range),
-                output_data_len_range: opts.data_len_range.map(convert_range),
-                output_capacity_range: opts.capacity_range.map(convert_range),
-                block_range: opts.block_range.map(convert_range),
-            })
-        };
-        SearchKey {
-            script: opts.primary_script.into(),
-            script_type: opts.primary_type.into(),
-            with_data: opts.with_data,
-            filter,
-            group_by_transaction: None,
-        }
-    }
 }
 
 #[derive(Serialize, Deserialize, Eq, PartialEq, Debug)]
@@ -95,13 +55,6 @@ pub struct TxWithCells {
     block_number: BlockNumber,
     tx_index: Uint32,
     cells: Vec<(CellType, Uint32)>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "snake_case")]
-pub enum CellType {
-    Input,
-    Output,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
